Move to legacy ok
